Remove Emojis
Hey Developer. This change was originally made for uYouEnhanced since a lot of people at that Repo doesn’t seem to like the emojis for some reason. Maybe there can be a reconsider about changing it back? Or possibly turning it into a toggle able option? That could work in my opinion. But anyways thanks again for your tweaks PoomSmart!