Better Fetching/Failed Format
Hey Developer. This change is currently exclusive to uYouEnhanced since a lot of them really don’t like the emojis in the Repo. Maybe you can reconsider about changing it back? Or possibly turning it into an option like I’m currently doing in this commit? That could work but not sure if it’s possible. But anyways thanks again for your tweaks PoomSmart! you are the best!